+++ src/Loader/index.js
@@ -4,7 +4,7 @@ const Template = require('./Template');
 const Loader = require('./Loader');
 const Dependency = require('./Dependency');
 const Options = require('./Options');
-const { preprocessorErrorToString, compileErrorToString, exportErrorToString } = require('./Messages/Exeptions');
+const { notInitializedPluginError, preprocessorError, compileError, exportError } = require('./Messages/Exeptions');
 
 const RenderMethod = require('./methods/RenderMethod');
 
@@ -19,16 +19,21 @@ const loader = function (content, map, meta) {
   const { loaderIndex, rootContext, resource, resourcePath } = loaderContext;
   let errorStage = 'init';
 
-  const callback = (error, result) => {
+  const callback = (error, result = null) => {
     if (error) {
       // in build mode, abort the compilation process and display an error in the output
-      if (!PluginService.isWatchMode()) return loaderCallback(error);
+      if (result == null || !PluginService.isWatchMode()) return loaderCallback(error);
 
       // in watch mode, display an error in the output without abort the compilation process
       loaderContext.emitError(error);
     }
     loaderCallback(null, result, map, meta);
   };
+
+  if (!PluginService.isUsed()) {
+    callback(notInitializedPluginError());
+    return;
+  }
 
   new Promise((resolve) => {
     // the options must be initialized before others
@@ -67,31 +72,30 @@ const loader = function (content, map, meta) {
     })
     .catch((error) => {
       const issuer = path.relative(rootContext, resourcePath);
-      let message;
 
       switch (errorStage) {
         case 'preprocessor':
-          message = preprocessorErrorToString(error, issuer);
+          error = preprocessorError(error, issuer);
           break;
         case 'compile':
-          message = compileErrorToString(error, issuer);
+          error = compileError(error, issuer);
           break;
         case 'export':
-          message = exportErrorToString(error, issuer);
+          error = exportError(error, issuer);
           break;
         default:
           // unrecoverable configuration error, Webpack restart required
           const method = new RenderMethod({});
-          const browserError = method.exportError(error, resource);
-          callback(error, browserError);
+          const browserErrorMessage = method.exportError(error, resource);
+          callback(error, browserErrorMessage);
           return;
       }
 
-      const browserError = Loader.exportError(message, resource);
+      const browserErrorMessage = Loader.exportError(error, resource);
 
       Dependency.init(loaderContext);
       Dependency.watch();
-      callback(message, browserError);
+      callback(error, browserErrorMessage);
     });
 };