+++ test/cases/loader-options-css-inline-auto-prod/webpack.config.js
@@ -1,0 +1,47 @@
+const path = require('path');
+const HtmlBundlerPlugin = require('../../../');
+
+module.exports = {
+  mode: 'production',
+
+  output: {
+    path: path.join(__dirname, 'dist/'),
+  },
+
+  resolve: {
+    alias: {
+      '@images': path.join(__dirname, 'src/assets/images/'),
+      '@styles': path.join(__dirname, 'src/assets/styles/'),
+    },
+  },
+
+  plugins: [
+    new HtmlBundlerPlugin({
+      entry: {
+        index: './src/views/index.html',
+        'pages/about': './src/views/about.html', // test same inline CSS from another path
+      },
+      css: {
+        inline: 'auto', // false; true; 'auto' - in dev mode is true, in prod mode is false
+        filename: 'assets/css/[name].[contenthash:8].css',
+      },
+    }),
+  ],
+
+  module: {
+    rules: [
+      {
+        test: /\.(css|sass|scss)$/,
+        use: ['css-loader', 'sass-loader'],
+      },
+
+      {
+        test: /\.(png|jpg)$/,
+        type: 'asset/resource',
+        generator: {
+          filename: 'assets/img/[name].[hash:8][ext]',
+        },
+      },
+    ],
+  },
+};