+++ test/cases/loader-options-js-inline-auto-prod/src/assets/scripts/script-inline-true.js
@@ -1,0 +1,11 @@
+const image = require('@images/apple.png');
+
+document.addEventListener('DOMContentLoaded', (event) => {
+  const elm = document.querySelector('.js-image-script-inline-true');
+  elm.style.backgroundImage = `url(${image})`;
+  elm.style.width = '160px';
+  elm.style.height = '130px';
+  elm.style.border = '2px solid orangered';
+});
+
+console.log('>> script-inline-true.js?inline=true', { image });